Minor refactor for ClientTests.
