--- java/integTest/src/test/java/glide/standalone/ClientTest.java
@@ -1,28 +1,0 @@
-/** Copyright GLIDE-for-Redis Project Contributors - SPDX Identifier: Apache-2.0 */
-package glide.standalone;
-
-import static glide.TestConfiguration.STANDALONE_PORTS;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import glide.api.RedisClient;
-import glide.api.models.configuration.NodeAddress;
-import glide.api.models.configuration.RedisClientConfiguration;
-import lombok.SneakyThrows;
-import org.junit.jupiter.api.Test;
-
-public class ClientTest {
-    @Test
-    @SneakyThrows
-    public void custom_command_info() {
-        RedisClient client =
-                RedisClient.CreateClient(
-                                RedisClientConfiguration.builder()
-                                        .address(NodeAddress.builder().port(STANDALONE_PORTS[0]).build())
-                                        .clientName("TEST_CLIENT_NAME")
-                                        .build())
-                        .get();
-
-        String clientInfo = (String) client.customCommand(new String[] {"CLIENT", "INFO"}).get();
-        assertTrue(clientInfo.contains("name=TEST_CLIENT_NAME"));
-    }
-}
