+++ java/integTest/src/test/java/glide/standalone/ClientTests.java
@@ -16,6 +16,21 @@ import org.junit.jupiter.api.Test;
 public class ClientTests {
     @Test
     @SneakyThrows
+    public void custom_command_info() {
+        RedisClient client =
+                RedisClient.CreateClient(
+                                RedisClientConfiguration.builder()
+                                        .address(NodeAddress.builder().port(STANDALONE_PORTS[0]).build())
+                                        .clientName("TEST_CLIENT_NAME")
+                                        .build())
+                        .get();
+
+        String clientInfo = (String) client.customCommand(new String[] {"CLIENT", "INFO"}).get();
+        assertTrue(clientInfo.contains("name=TEST_CLIENT_NAME"));
+    }
+
+    @Test
+    @SneakyThrows
     public void close_client_throws_ExecutionException_with_ClosingException_cause() {
         RedisClient client =
                 RedisClient.CreateClient(
